+++ engine/src/xetex_xetexd.rs
@@ -8,20 +8,20 @@ use crate::{xetex_ini, xetex_output};
  * this conversion besides painstakingly annotating things.
  */
 
-pub(crate) unsafe fn LLIST_link(p: isize) -> &'static mut i32 {
-    &mut MEM[p as usize].b32.s1
+pub(crate) unsafe fn LLIST_link(p: usize) -> &'static mut i32 {
+    &mut MEM[p].b32.s1
 }
-pub(crate) unsafe fn LLIST_info(p: isize) -> &'static mut i32 {
-    &mut MEM[p as usize].b32.s0
+pub(crate) unsafe fn LLIST_info(p: usize) -> &'static mut i32 {
+    &mut MEM[p].b32.s0
 }
 
 /// half of LLIST_info(p)
-pub(crate) unsafe fn NODE_type(p: isize) -> &'static mut u16 {
-    &mut MEM[p as usize].b16.s1
+pub(crate) unsafe fn NODE_type(p: usize) -> &'static mut u16 {
+    &mut MEM[p].b16.s1
 }
 /// the other half of LLIST_info(p)
-pub(crate) unsafe fn NODE_subtype(p: isize) -> &'static mut u16 {
-    &mut MEM[p as usize].b16.s0
+pub(crate) unsafe fn NODE_subtype(p: usize) -> &'static mut u16 {
+    &mut MEM[p].b16.s0
 }
 
 /// subtype; records L/R direction mode
